requirements.txt
pytest==8.4.2
selene==2.0.0rc9
selenium==4.38.0
allure-pytest==2.15.3
allure-python-commons==2.15.3